+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to qauvern will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2026-06-xx
+
+## Changed
+
+- `qauvern` now eagerly errors if the config file has any archived instances.
+
+## Fixed
+
+- `qauvern configure` no longer includes archived instances.
+
 ## [0.4.0] - 2026-05-28
 
 ### Changed